+++ scripts/prepare_sector_network.py
@@ -4775,6 +4775,7 @@ def add_offshore_grid_tyndp(
     pyear: int,
     offshore_grid_fn: str,
     costs: pd.DataFrame,
+    options: dict,
     nyears: float = 1,
 ):
     """
@@ -4792,6 +4793,9 @@ def add_offshore_grid_tyndp(
         Path to the file containing offshore grid configuration data.
     costs : pd.DataFrame
         Technology costs assumptions.
+    options : dict
+        Configuration options containing at least:
+        - offshore_hubs.connect_isolated : bool
     nyears : float, default 1
         Number of years for which to scale the investment costs.
 
@@ -4864,6 +4868,38 @@ def add_offshore_grid_tyndp(
         lifetime=costs.at["H2 (g) submarine pipeline", "lifetime"],
     )
 
+    # Copperplate isolated wind farms to the grid
+    if options["offshore_hubs_tyndp"]["connect_isolated"]:
+        links_oh = n.links.query("carrier.isin(['DC_OH', 'H2 pipeline OH'])")
+        buses_oh_idx = n.buses.query("carrier in ['AC_OH', 'H2_OH']").index
+
+        query_str = "p_nom_extendable or p_nom > 0"
+        links_oh_cap = links_oh.query(query_str)
+        links_oh_no = links_oh.query(f"~({query_str})")
+
+        buses_cap = set(pd.concat([links_oh_cap.bus0, links_oh_cap.bus1])).intersection(
+            buses_oh_idx
+        )
+        buses_no = set(pd.concat([links_oh_no.bus0, links_oh_no.bus1])).intersection(
+            buses_oh_idx
+        )
+        buses_unconnected = set(buses_no) - set(buses_cap)  # noqa: F841
+        buses_mainland = n.buses.query(  # noqa: F841
+            "carrier in ['AC', 'H2', 'H2 Z1', 'H2 Z2']"
+        ).index
+        buses_target = n.generators.query(  # noqa: F841
+            "carrier.str.contains('offwind') "
+            "and bus in @buses_unconnected "
+            "and p_nom > 0"
+        ).bus
+
+        idx_patch = links_oh_no.query(
+            "bus0 in @buses_target and bus1 in @buses_mainland "
+        ).index
+
+        n.links.loc[idx_patch, "p_nom"] = np.inf
+        n.links.loc[idx_patch, "lifetime"] = np.inf
+
 
 def add_offshore_hubs_tyndp(
     n: pypsa.Network,
@@ -4874,6 +4910,7 @@ def add_offshore_hubs_tyndp(
     profiles_pecd: pd.Series,
     costs: pd.DataFrame,
     spatial: SimpleNamespace,
+    options: dict,
     nyears: float = 1,
 ):
     """
@@ -4900,7 +4937,10 @@ def add_offshore_hubs_tyndp(
         Technology costs assumptions.
     spatial : object, optional
         Object containing spatial information about nodes and their locations.
-    nyears : float
+    options : dict
+        Configuration options containing at least:
+        - offshore_hubs.connect_isolated : bool
+    nyears : float (default : 1)
         Number of years for which to scale the investment costs.
 
     Returns
@@ -4950,7 +4990,7 @@ def add_offshore_hubs_tyndp(
     add_offshore_electrolysers_tyndp(n, pyear, offshore_electrolysers_fn, costs, nyears)
 
     # Add offshore DC and H2 grid connections
-    add_offshore_grid_tyndp(n, pyear, offshore_grid_fn, costs, nyears)
+    add_offshore_grid_tyndp(n, pyear, offshore_grid_fn, costs, options, nyears)
 
 
 def attach_gas_load(
@@ -9617,6 +9657,7 @@ if __name__ == "__main__":
             profiles_pecd=profiles_pecd,
             costs=costs,
             spatial=spatial,
+            options=options,
             nyears=nyears,
         )